requirements.txt
alembic==1.3.2
appnope==0.1.0
Babel==2.7.0
backcall==0.1.0
certifi==2019.11.28
Click==7.0
decorator==4.4.1
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
ipython==7.10.2
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.17.4
pandas==0.25.3
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.2
psycopg2==2.8.4
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.11
traitlets==4.3.3
wcwidth==0.1.7
Werkzeug==0.16.0
WTForms==2.2.1
yapf==0.28.0